requirements.txt
absl-py==0.1.12
astor==0.6.2
bleach==1.5.0
click==6.7
cycler==0.10.0
decorator==4.3.0
Django==2.0.3
Flask==0.12.2
Flask-Cors==3.0.3
gast==0.2.0
grpcio==1.10.0
html5lib==0.9999999
image==1.5.20
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
networkx==2.1
nltk==3.2.5
numpy==1.14.2
opencv-python==3.4.0.12
Pillow==5.0.0
protobuf==3.5.2.post1
pymongo==3.6.1
pyocr==0.5.1
pyparsing==2.2.0
python-dateutil==2.7.0
pytz==2018.3
PyWavelets==0.5.2
scikit-image==0.13.1
scipy==1.0.1
six==1.11.0
termcolor==1.1.0
Wand==0.4.4
Werkzeug==0.14.1
